pyproject.toml
[project]
name = "disgrok"
version = "0.1.0"
description = "Discord bot with Grog integration"
authors = [
    {name = "Kurt Andrews"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp (>=3.10,<3.10.11)",
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
